pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chex-sae-fairness"
version = "0.1.0"
description = "SAE interpretability and fairness auditing pipeline for CheXagent + CheXpert Plus"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Research Team" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.12",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "torchvision>=0.17",
  "transformers>=4.39",
  "Pillow>=10.2",
  "pyyaml>=6.0.1",
  "tqdm>=4.66",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.3",
]

[project.scripts]
chex-prepare-manifest = "chex_sae_fairness.cli:prepare_manifest_cli"
chex-extract-features = "chex_sae_fairness.cli:extract_features_cli"
chex-train-sae = "chex_sae_fairness.cli:train_sae_cli"
chex-run-study = "chex_sae_fairness.cli:run_study_cli"
chex-run-sae-sweep = "chex_sae_fairness.cli:run_sae_sweep_cli"
chex-audit-data = "chex_sae_fairness.cli:audit_data_cli"
chex-run-core-paper = "chex_sae_fairness.cli:run_core_publication_cli"
chex-run-supplement-paper = "chex_sae_fairness.cli:run_supplement_publication_cli"
chex-init-paper-config = "chex_sae_fairness.cli:init_publication_config_cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
